CLI cleanup
