+++ cmd/agent/api/agent/agent.go
@@ -14,6 +14,7 @@ import (
 
 	log "github.com/cihub/seelog"
 
+	"github.com/DataDog/datadog-agent/cmd/agent/common"
 	"github.com/DataDog/datadog-agent/pkg/flare"
 	"github.com/DataDog/datadog-agent/pkg/status"
 	"github.com/DataDog/datadog-agent/pkg/util"
@@ -50,7 +51,7 @@ func getHostname(w http.ResponseWriter, r *http.Request) {
 
 func makeFlare(w http.ResponseWriter, r *http.Request) {
 	log.Infof("Making a flare")
-	filePath, err := flare.CreateArchive(false)
+	filePath, err := flare.CreateArchive(false, common.GetDistPath(), common.PyChecksPath)
 	if err != nil || filePath == "" {
 		if err != nil {
 			log.Errorf("The flare failed to be created: %s", err)
