+++ cmd/agent/app/listchecks.go
@@ -13,20 +13,20 @@ import (
 )
 
 func init() {
-	AgentCmd.AddCommand(listCheckCommand)
-
+	// TODO: re-enable when the API endpoint is implemented
+	// AgentCmd.AddCommand(listCheckCommand)
 }
 
 var listCheckCommand = &cobra.Command{
-	Use:   "listchecks",
-	Short: "Query the running agent for the hostname.",
+	Use:   "list-checks",
+	Short: "Query the agent for the list of checks running",
 	Long:  ``,
 	RunE:  doListChecks,
 }
 
 // query for the version
 func doListChecks(cmd *cobra.Command, args []string) error {
-	c := GetClient()
+	c := common.GetClient()
 	urlstr := "http://" + sockname + "/check/"
 
 	body, e := common.DoGet(c, urlstr)
