+++ cmd/agent/common/getclient_windows.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2017 Datadog, Inc.
 
-package app
+package common
 
 import (
 	"net"
